Added main.h, utils.h to further refactor the code base, moved hide/protect toggles inside stealth_helper.h and done minor modifications in utils.h
